Starting File
I copied in the code we are going to reuse from last year. I copied in modify2 and initiated updateWheels but didn't write the code for it. updateWheels() needs to be made by the end of the week so Lucas can test it over the break. If you want to do it or are about to start it, let everyone know in the discord so two different people don't do it at the same time on accident.