First updateLauncher ()
updateLauncher() first draft. When button 8R is held down, the launcher turns on. Otherwise, the launcher stays off. However, it goes straight from still to max speed which will damage the motors and if a motor were to get damaged during the competition, the amount of time required to change it would put us out of at least one match. The ports need to be changed to matched to what is plugged into the robot. This should be its own branch but I forgot to make one before I edited this.