It Runs Now
Put the while loop in the main task instead of in each mechanism's task. Changed conveyor so when you push down the button, it turns on. Rather than press the button to turn it on and press the button to turn it off. Switched intake and conveyor ports so because conveyor was in port 1.