Combined Autonomous and User Controlled
Combined the autonomous and user control programs into one file. Will be tested during 7th period today to ensure it works. A copy of the original working file is saved just in case it does not work as hoped. Deleted any methods that aren't being used to make it cleaner and easier to read and make changes to.